delete and save not working.. next up

+++ BookStore/src/main/java/com/example/BookStore/BookRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.repository.CrudRepository;
 import javax.transaction.Transactional;
 import java.util.List;
 
-public interface BookRepository extends CrudRepository<Book, Long> {
+public interface BookRepository extends CrudRepository<Book, Integer> {
 
     List<Book> findAllByOrderByAuthor();
     List<Book> findAllByOrderByAuthorDesc();
